backend/btc: move SigScriptWitnessSize to txsize.go
More fitting place, and only called from that package too.
